+++ Sources/KeyboardKit/Layout/View+KeyboardLayoutWidth.swift
@@ -1,5 +1,5 @@
 //
-//  View+KeyboardButtonWidth.swift
+//  View+KeyboardLayoutWidth.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-01-21.
@@ -14,19 +14,22 @@ public extension View {
      Apply a certain keybard button width type to this view.
      
      The `totalWidth` parame is only required when using the
-     `.percentage` and `.reference(.percentage)` width types.     
+     `.percentage` and `.reference(.percentage)` width types.
      
      The `referenceSize` binding param is only required when
      using `.reference` and `.fromReference` width types.
+     
+     `TODO` Unit test these extensions.
      */
     @ViewBuilder
-    func width(_ width: KeyboardButtonWidth, totalWidth: CGFloat = .zero, referenceSize: Binding<CGSize> = .constant(.zero)) -> some View {
+    func width(_ width: KeyboardLayoutWidth, totalWidth: CGFloat = .zero, referenceSize: Binding<CGSize> = .constant(.zero)) -> some View {
         switch width {
-        case .fromReference: self.frame(width: referenceSize.width.wrappedValue)
+        case .available: self.frame(maxWidth: .infinity)
         case .percentage(let percent): self.frame(width: percent * totalWidth)
         case .points(let points): self.frame(width: points)
-        case .reference(let width): self.frame(width: self.width(for: width, totalWidth: totalWidth, referenceSize: referenceSize.wrappedValue)).bindSize(to: referenceSize)
-        case .available: self
+        case .reference(let width): self.frame(maxWidth: .infinity).frame(width: self.width(for: width, totalWidth: totalWidth, referenceSize: referenceSize.wrappedValue)).bindSize(to: referenceSize)
+        case .useReference: self.frame(width: referenceSize.width.wrappedValue)
+        case .useReferencePercentage(let percent): self.frame(width: percent * referenceSize.width.wrappedValue)
         }
     }
 }
@@ -36,11 +39,11 @@ private extension View {
     /**
      Calculate the width in points for a certain width type.
      */
-    func width(for width: KeyboardButtonWidth, totalWidth: CGFloat, referenceSize: CGSize) -> CGFloat? {
+    func width(for width: KeyboardLayoutWidth, totalWidth: CGFloat, referenceSize: CGSize) -> CGFloat? {
         switch width {
+        case .available: return nil
         case .percentage(let percent): return percent * totalWidth
         case .points(let points): return points
-        case .available: return nil
         default: return 0
         }
     }