+++ RELEASE_NOTES.md
@@ -31,6 +31,7 @@ Besides the new things listed below, there are a bunch of new extensions, images
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputActionProvider`.
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputCalloutContext`.
 * `KeyboardInputViewController` `setup(with:)` is now open and overridable.
+* `Layout` has new types for working with keyboard layouts.
 * `Locale` is a new namespace with a few new locale-specific utils.
 * `LocaleDictionary` is a simple dictionary wrapper to work with localized resources.
 * `LocaleKey` is a simple enum to gather top-level locale identifiers.
@@ -74,6 +75,7 @@ Besides the new things listed below, there are a bunch of new extensions, images
 * `KeyboardAppearance` button properties are prefixed with `button`.
 * `KeyboardAppearance` `font` and `fontWeight` have been merged into a single `font` property.
 * `KeyboardAppearanceProvider` has been renamed to `KeyboardAppearance`.
+* `KeyboardButtonWidth` has been renamed to `KeyboardLayoutWidth`.
 * `KeyboardContext` `actionHandler` has been moved to the input vc.
 * `KeyboardContext` `keyboardAppearanceProvider` has been moved to the input vc.
 * `KeyboardContext` `keyboardBehavior` has been moved to the input vc.